Remove feature toggle and adjust tests
