+++ src/sql/QueryEditor/FormatSelect.tsx
@@ -3,7 +3,6 @@ import { DataQuery, SelectableValue } from '@grafana/data';
 import { InlineField, Select } from '@grafana/ui';
 
 export type FormatSelectProps<TQuery extends DataQuery, FormatOptions> = {
-  newFormStylingEnabled?: boolean;
   id?: string;
   query: TQuery;
   options: Array<SelectableValue<FormatOptions>>;
@@ -21,10 +20,7 @@ export function FormatSelect<TQuery extends DataQuery & Record<string, any>, For
     });
     props.onRunQuery?.();
   };
-  return (
-    <>
-      {props.newFormStylingEnabled ? (
-        <Select
+  return <Select
           aria-label="Format data frames as"
           id={props.id ?? 'formatAs'}
           options={props.options}
@@ -32,18 +28,4 @@ export function FormatSelect<TQuery extends DataQuery & Record<string, any>, For
           onChange={onChangeFormat}
           menuShouldPortal={true}
         />
-      ) : (
-        <InlineField label="Format as" labelWidth={11}>
-          <Select
-            aria-label="Format as"
-            options={props.options}
-            value={props.query.format}
-            onChange={onChangeFormat}
-            className="width-12"
-            menuShouldPortal={true}
-          />
-        </InlineField>
-      )}
-    </>
-  );
 }
